broker: add topology_get_internal_ranks()
Problem: There is no way to get the set of internal ranks of a TBON
topology, but this will be useful when determining which broker ranks
are critical to a running instance.

Add topology_get_internal_ranks() to get the set of internal nodes in
the current broker topology. Add unit tests to verify the function
operation.