job-exec: add a critical-ranks RPC
Problem: It would be useful if jobs could notify their parent instance
of the set of job shell ranks (i.e. nodeids) that should generate a
fatal job exception when they are lost or fail, but no such feature
exists.

Add a critical_ranks idset to the job-exec jobinfo structure which
is initialized to the set of all job shell ranks (i.e. zero origin
to size - 1). Add a job-exec.critical-ranks RPC via which a job can
update this set of ranks.

For now, do nothing with this new jobinfo member.