job-exec: only raise fatal exception if lost rank is critical
Problem: The job-exec system generates a fatal exception any
time the connection to a job shell is "lost", e.g. node or broker
failure. However, a fatal exception should only be generated if the
lost shell rank is in the job's set of critical ranks.

Raise a non-fatal exception and notify the rank 0 job shell of the lost
shell via the shell.exception RPC when the lost rank is not in the set
of critical ranks. By default, all ranks are in the set of critical
ranks unless the set is updated via the job-exec.critical-ranks RPC,
so there should be no change in behavior when the critical_ranks
attribute is unchanged.

Fixes #4573