testsuite: test resiliency of batch jobs
Problem: There is no test in the testsuite that ensures instances of
Flux run as jobs can tolerate the loss of non-critical ranks.

Add a test to the issues testsuite that ensures a simple instance
survives the loss of a leaf broker and can run a job on the remaining
broker ranks.