job-info: return error on invalid eventlog append
Problem: If an eventlog is illegally overwritten, it can result
in the appearance of a "zero length" append.  If a user is watching
this eventlog, this "zero length" append can result in a segfault
in the job-info module.

Solution: If a "zero length" append occurs in an eventlog being
watched, consider the eventlog malformed and return an error.

Fixes #4612