job-manager: add queue class
Problem: there is no way to list queues or enable/disable job
submission on a queue basis.

Add a new class to the job manager which provides new interfaces
for enabling, disabling, listing, and querying status of queues.

Note that these queues are not containers for jobs.  Jobs are still
enqueued in one "alloc queue" even when there are multiple named
queues configured.  The purpose of this class (at this time) is to
capture the queue configuration and the administrative status to be
be utilized by the job submit logic.

A future commit will wire this into the job submit logic.