python: move job-list fields to attrs into bindings
Problem: The translation of JobInfo properties/fields to job-list
attrs is done within a function in flux-jobs.py and thus is not
available to other users of the Python job listing and formatting
APIs. Furthermore, when adding a new property to JobInfo for use as
a field in flux-jobs(1), the new property must be updated in several
places (including flux-jobs.py), which is somewhat annoying and
error prone.

Move the mapping of job fields to job-list attributes into the
Python bindings as flux.job.job_fields_to_attrs(). This not only
allows access to this mapping from users other than flux-jobs(1),
but centralizes most of the places where a new job field needs to be
added to src/bindings/python/flux/job/info.py.