python: add contextual_info JobInfo property
Problem: It woudl be convenient to have a JobInfo property which would
return some useful information for pending jobs in place of the nodelist
attribute, but there is not a convenient way to currently do that.

Add a 'contextual_info' property to the JobInfo class. This property will
return a different bit of data depnding on the job state, i.e.

 - In PRIORITY, the string "priority-wait" is returned
 - In DEPEND, "depends:" + the list of outstanding dependencies is returned
 - In SCHED, "eta:" + the remaining time until the scheduler's t_estimate
   (if available) is returned
 - O/w, the nodelist is returned

This allows `{contextual_info}` to be used in flux-jobs(1) output so that
interesting data is perhaps shown for pending jobs.