testsuite: add issues test driver to LONGTESTSCRIPTS
Problem: The t4000-issues-test-driver.t test is one of the longer
tests, but runs near the end of the testsuite, which extends the
overall time to run `make -j N check`.

Add t4000-issues-test-driver.t to LONGTESTSCRIPTS which are run near
the beginning of the testsuite in an attempt to decrease the overall
testsuite time.