flux-jobs: fix possible UnicodeEncodeError on output
Problem: If flux-jobs is running in a non-utf8 locale and tries to
render the dependencies field, some lines may contain jobids with
the unicode `ƒ` which causes flux-jobs to raise a UnicodeEncodeError
exception.

If a UnicodeEncodeError exception is raised, then encode the offending
line as utf-8 with errors=surrogateescape to avoid the error.