broker: reduce default tbon.tcp_user_timeout
Problem: the default time to wait for a TCP ACK from a peer
broker before disconnecting is too long.  The system default
is around 20m, which is a very long time for a free range
batch job to be hung up after losing a non-critical node.

Use 20s as the default.  This can still be overridden via TOML
config and broker command line.

Update sharness test that checked for a default of 0 (system).