consolidate python linting and formatting
Problem: the developer experience is challenging because linting
and formatting are hard to debug locally, and then the separation
in the CI means 3+ clicks to see what is going on!

To fix this, this changeset adds the use of pre-commit, which
will always run linting/formatting (and with automated fixes for
a subset of the tools) right before commit, so the developer can
have some confidence that passing locally == passing in the CI. In
the case that there is a bug in the CI, we will now just have one
place to look instead of 3+. This should be a start of a new setup,
including addition of pre-commit and consolidation of settings into
a pyproject.toml (and setup.py for flake8) and we can further tweak
as we establish our linting preferences.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>