flux-uptime: use new queue-status RPC
Problem: flux-uptime uses the deprecated job-manager.submit-admin RPC.

Use the job-manager.queue-status RPC instead.

Since the purpose of this section of code is to add a warning to the
output when the queue is disabled, and not to provide comprehensive queue
status, the warning is now added only when all queues are disabled.