flux-queue: support multiple queues
Problem: there is no way to list available queues, nor enable/disable
individual queues.

Add a -q,--queue option to the enable, disable, and status subcommands.
If the queue option is unspecified and multiple named queues are configured,
all queues are targetted (--all confirmation required for enable, disable).
This allows the commands to function similar to the way they did before when
there is only the anonymous queue, which is still the common case.

Since flux queue status now lists the enable/disable status of all queues
when none are specified, this is now a way for a user to find out what
queues are available on the system.

Update rc1 to use "flux queue disable --all" when putting the instance
in safe mode.

Fixes #4620