job-manager: integrate queue class with submit logic
Problem: jobs are accepted when submitted with a named queue when
queues are not configured.

Call queue_submit_check() on each submitted job.  The job is rejected
if a requested queue is not configured, or if the queue is disabled.

As part of the integration, the job-manager.submit-admin RPC is moved
temporarily to queue.c.  It will be removed once tools have been updated
to use the new RPCs.

Fixes #4440