python: add contextual_time property to JobInfo class
Problem: It would be useful to allow a column in flux-jobs output
which displays duration (if set) for pending jobs, and the runtime
for jobs that reach the RUN state, but no such attribute exists in
the JobInfo class.

Add a `contextual_time` attribute, which displays the job duration
for pending jobs (or jobs that never reached the RUN state), otherwise
the job runtime.