job-list: skip unnecessary cleanup path
Problem: In check_id_valid_continuation(), if we need to continue
to wait for an id to be valid, we go through the function's cleanup
path.  This is not necessary as the cleanup is basically a no-op.

Solution: Do not go down the cleanup path is job is required to
wait further.  Instead just return.