job-manager: raise fatal error when configured plugin fails to load
Problem: When a jobtap plugin that is loaded via job-manager
configuration fails to load for any reason, the error is ignored and
the instance proceeds normally. This could lead to later failures
that are much more difficult to diagnose.

The test for failure of jobtap_load_plugin() in jobtap_conf_entry()
tests for a result `< 0` when a check for NULL should be used instead.

Fix the test so that failures to load plugins from the config file
are detected by the instance. The result is now a hard failure.

Fixes #4645