testsuite: ensure job-manager.plugins load failure is detected
Problem: No tests in the testsuite ensure that a jobtap plugin that
is required by configured and fails to load is detected and causes
a fatal error to the starting instance.

Add a load error test to t2212-job-manager-plugins.t that ensures a
missing plugin that is required by configuration causes a fatal error.

Co-authored-by: Jim Garlick <garlick.jim@gmail.com>