doc: add note about flux --parent option in flux-mini(1)
Problem: A common use case is for a user to submit a new batch
script to the parent instance from within a batch job, but the
flux-mini(1) man page doesn't mention how to do this.

Add a note to the flux-mini(1) man page about the flux(1) `--parent`
option, which can be used to target the parent instance for any
flux command, but is useful here to submit a new batch job from
a batch script.