broker: launch non-interactive shells in a new process group
Problem: When the broker launches a non-interactive shell via runat,
e.g. when running a script as part of a batch job, and that script
is later canceled, the signal is delivered only to the shell which
may leave other processes running. This can cause batch jobs to hang
after a `flux shutdown` if the batch script is running a command in the
foreground. The shell will be terminated, but the command keeps running
keeping the stdio file descriptors open so the job shell can't exit.

When the runat subsystem is spawning a shell command, add the
subprocess flag FLUX_SUBPROCESS_FLAGS_SETPGRP, so that the shell is
run in a new process group. This then causes flux_subprocess_kill(3)
to call killpg(2) instead of only signaling the spawned shell.

Fixes #4596