support driving a handle from external event loop
This interface mirrors the one used by libuv, which offers:
* uv_backend_fd
* uv_backend_timeout

For flux this adds flux_reactor_resume_timeout, which when combined with
flux_pollfd and some extra timer logic can be used to drive a handle
to completion with relatively efficient blocking behavior.  The design
is that the external event loop should set up watchers much like ev_flux
does to handle IO events, but additionally in the check callback
should invoke flux_reactor_resume_timeout to determine what blocking
mode flux expects.  The logic is:
* If the reactor has a prepare, check, or idle watcher *currently
  active* then it will return 0, and expect the outer loop to call back
  into it on the next iteration of the outer loop.  The outer loop
  should use an idle watcher or similar to ensure this makes progress.
* If the reactor has either timer or periodic watchers registered, then
  it will return the minimum registered relative timeout or pending
  fire time.  The outer loop is expected to call back into the flux
  reactor at that timeout, and possibly before.  Using a timer or
  periodic call into flux should both work.
* If the reactor finds neither pending timeouts nor per-iteration
  watchers, then it will return -1, indicating that the outer loop
  can block on the pollfd of the handle being embedded until an IO
  event arrives.

For the case of embedding a handle's pollfd and associated continuations
and reactor events into an external loop this appears to work well.  It
does not handle *all* reactor events well necessarily.  In particular
signal and child watchers are suspect, and should not be used with this
interface.  The IO, per-iteration and time event watchers should all be
safe.