job-ingest/test: skip guest tests when default sign-type fails
Problem: The job-ingest test_job.t unit test fails if the build
environment is not able to sign jobspec -- e.g. if sign-type=munge is
configured but the munge daemon is not running. But the purpose of
this unit test is not to test flux-security but to test job-ingest
interfaces, so forcing failure due to misconfigured environment is
not necessary.

Just skip the "guest" tests when the default sign-type is not
functional.