testsuite: rename variables to avoid confusion
Problem: In t2260-job-list.t, local variables "nnodes" and
"ncores" are used to store the total nodes and total cores
in the flux instance.  This can be confusing because "nnodes"
and "ncores" are also fields returned by job-list, which is tested
throughout t2260-job-list.t.

Rename the variables to "totalnodes" and "totalcores".