broker/test: Fix runat race on older glibc versions
Problem: On older glibc versions, the setting of the child process
group via POSIX_SPAWN_SETPGROUP can be racy.  A call to killpg()
immediately after a spawned process is launched may not kill
a child process.  In one runat test, this can give the appearance
of a hang as a 3600 second sleep call is being waited on.

Solution: On older glibc versions, for the use of fork/exec over
posix_spawn for this particular racy test.

Fixes #4657