completions: add multi-level bash completions file
Problem: The existing bash completion script could only complete
first-level flux(1) subcommands, and also disabled filename completion
for commands where it was useful like `flux mini batch`.

Rewrite the bash completion script, including expanded support for
many flux multi-level subcommands like `flux mini run,batch,..`, `flux
resource list,status,..` and even `flux kvs eventlog get,put` etc.

Filename completion is retained whenever there are not specific
completions provided by the script. This won't be applicable to all
flux commands, but this allows filename completion to work where it
is helpful, e.g.  `flux mini batch` etc.

Not all commands are represented, but this is a good start for the
most useful and longest flux commands.

Additionally, experimental support for targeted completions are
included such as

 - active jobid completion for `flux job` subcommands where it makes
   sense (e.g. `flux job cancel <TAB>`)
 - queue name completion for the `-q, --queue=` options in `flux mini`
   `flux queue` and `flux jobs`.
 - loaded module name completion in flux-module(1) commands like
   remove, reload, info, and stats
 - broker attribute completion for flux-getattr(1)

Many of the command options are hard-coded here. This is unfortunate
but automating the gathering of subcommands and options is left as
an exercise for the future.

Fixes #4457