build: add get_builtins.sh as prereq for completions
Problem: The get_builtins.sh script is not listed as a prerequisite
of etc/completions/flux, though it is used in building that file.

Add get_bultins.sh as a prereq for completions/flux. This allows
the completions to be regenerated if get_builtins.sh is edited.

Also, drop the unnecessary use of $(srcdir) here, which is implied.