rc1: reload configuration on rank > 0
Problem: although flux shutdown stops the rank 0 broker in a system
instance, the other ranks are immediately restarted by systemd.
This means that altering the configuration while flux is "down"
may be too late for the brokers that have already restarted but have
not yet rejoined the instance.

To avoid confusing people, run flux config reload in rc1 on
rank > 0 to ensure it starts with the latest configuration.

Fixes #4663