python: change return type of OutputFormat.filter_empty
Problem: The OutputFormat.filter_empty() method returns a new instance
of the class, but this makes it difficult to use this function from
subclasses of OutputFormat that have different constructor arguments.

Change the filter_empty() method to return the new format string,
so that the caller can construct the result with the appropriate
constructor.