python: save original format string in OutputFormat class
Problem: The original provided format is lost after the OutputFormat
constructor if the prepend option is used. This means the filter_empty()
method returns the prepended format string, which can cause errors
where the prepended format fields are not expected.

Save the original format string in OutputFormat and add an orig
parameter that, when true, returns the original string instead of the
modified format string.