broker: add 'custom' topology plugin
Problem: overlay topology is limited to complete k-ary, but a system
instance may require the topology to be mapped to the cluster architecture,
for example, using service nodes as TBON routers.

Add a "custom" topology plugin that scans the bootstrap hosts array
for "parent" directives.  Rank 0 is the default parent of ranks > 0
that do not explicitly specify a parent.

Example:

hosts = [
    { host = "test1" },
    { host = "test[0,64,128,192]", parent = "test1" },
    { host = "test[2-63]",    parent = "test0" },
    { host = "test[65-127]",  parent = "test64" },
    { host = "test[129-191]", parent = "test128" },
    { host = "test[193-255]", parent = "test192" }
]

Update tests.

Fixes #3799