python: allow `+` suffix in UtilFormatter to denote truncation
Problem: It is common for formatted output created by the UtilFormatter
class and its subclasses to truncate fields, but there is no inherit
way to indicate this truncation in the result, which can cause
confusion.

Support a `+` suffix for format specs in UtilFormatter that, when
present, causes the string result from format_field() to get a
truncation indicator character (currently "+") if it appears there
was truncation due to the format.

If both the "+" and "h" (hyphenation) suffix characters are used, then
the "+" must appear last in the format spec, e.g. `{field:>6.6h+}`.