job-list: add queue specific stats
Problem: The job-list module tracks stats for all jobs.  It would
be useful to track stats for individual queues as well.

Solution: Add queue stats tracking to the internal stats library
as well.  The stats of each queue will be returned in the 'queues'
array of the object returned by `job-list.job-stats'.

Fixes #4604