add devcontainer environment for vscode
Problem: it is hard to develop flux-core locally.
This addition will allow VSCode users to quickly
run (and work in) a development container, defined
under .devcontainer in the root, which contains a
devcontainer.json file and a Dockerfile that uses
the fluxrm/testenv:focal base to quickly pull and
spin up the container. We need the Dockerfile to
support a post start command to fix a git security
issue, and also add developer tools like formatters,
linters, etc. This is a vanilla install, meaning it
just will allow for build of the core flux, and as
other developers try this out we can easily add
other extensions that make the development experience
better!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>