flux-mini: support extended syntax with --requires
Problem: The --requires option in flux-mini only accepts a list
of properties, so there is no way for users to specify even basic
boolean expressions (such as "xx or yy") or other operators (such as
the hostlist and ranks operators) in job constraints.

Add a Constraint class to flux.job with a simple parser that converts
a boolean expression like:

 xx and yy or (hosts:foo[0-1] and gpu)

to an RFC 31 Constraint object. Generic operators are separated
from their arguments with OP:ARGS, if no operator is given, the
"properties" operator is assumed. Boolean operators "and" "or" and
"not" are supported, as are parentheses for enforcing precedence. If a
generic operator is preceeded by a `-`, then that operator is joined as
the right hand side of a "not". For ease of use, "hosts" and "host"
are treated as synonyms for "hostlist" and "rank" is treated as a
synonym for "ranks".

Replace the existing flux-mini --requires handler using Constraint().