broker: allow duplicate hosts entries
Problem: bootstrap.hosts handles hosts array entries with the same host key
by assigning one host multiple ranks, which is incorrect.

When entries with the same host key are encountered, combine them, retaining
the rank assigned to the first one.  This allows host entries to be repeated
for useful effect, like the following, which preserves a numerical mapping
of ranks to hostname suffixes, while allowing a custom topology to be
expressed in duplicate entries.

 [[bootstrap.hosts]]
 host = "test[0-127]"
 [[bootstrap.hosts]]
 host = "test[1,64]"
 parent = "test0"
 [[bootstrap.hosts]]
 host = "test[2-63]"
 parent = "test1"
 [[bootstrap.hosts]]
 host = "test[65-127]"
 parent = "test64"

Update unit test.