flux-resource: add --skip-empty option for status command
Problem: The flux-resource status command skips empty lines of output
(lines where the number of nodes is zero) unless the verbose option
was used or states were explicitly requested. This is fine most of
the time, but there may be cases where a user wants to skip empty
lines even when these other two options are used.

Add a --skip-empty option which causes flux-resource status to always
skip empty lines.