flux-resource: implement drain_list() with status()
Problem: Now that `flux resource status` supports the {timestamp}
field, `flux resource drain` with no other arguments is just a call to
`flux resource status` with a specific set of arguments.

Update drain_list() (the function called when `flux resource drain`
is called with no arguments) to just call status() with a specific
format and set of options.