devtools: add script to generate clang comp. db
This is a thin wrapper around bear and make, but it should help generate
a compile_commands.json database for use with clangd, clang-tidy, or any
IDE or language server that can use one for more precise C/C++ handling.

Basic usage, after configuring:
./scripts/generate_compile_commands <make-args>

by default it will use the `all` target, to get everything add the `-t`
option to force it to use check so that it will include all the tests in
the database as well.