devenv: expand devcontainer.json and Dockerfile
This adds the necessary extensions and support tools to the devcontainer
to support using vscode with full functionality.  It also fixes a typo
in the original that caused git to flake out.

The additional tools also make it easy to generate a
compile_commands.json for use with other editors like vim with clangd.