devenv: base setup for vscode in flux
This adds recommended extensions, configure build and test tasks, and
settings to provide solid base c, python and lua language features for
flux if the project is opened in vscode.  It intentionally avoids doing
anything complex or requiring a large number of extensions, but should
work well in the base case.

The only little extra is I included the github pull-request interface in
the extensions, since using it for reviewing is quite useful and nice to
have by default.