python: update flux.util.OutputFormat constructor
Problem: The flux.util.OutputFormat class has a headings class
attribute, but the constructor still requires a positional
valid_headings parameter. This unnecessarily complicates subclasses
of the OutputFormat class. Also, all extant use cases of OutputFormat
use prepend="0." so this should just be the default.

Update the OutputFormat.__init__() function to take the format as the
only positional parameter along with optional headings and prepend
arguments. The prepend argument now defaults to "0.", which will
by far be the common case, and the headings class attribute is only
overridden if headings is not None.

Update all callers.