python: make deep copy of initial dict in UtilConfig
Problem: If the initial_dict passed to the UtilConfig constructor is
deeper than a single level, then the configuration object contains
a reference to the original dict, which may change as configuration
files are loaded and merged into self.dict.

Use copy.deepcopy() to create a deep copy of the initial_dict in the
UtilConfig constructor to avoid changes to the original.