libflux: support flux_get_process_scope()
Problem: It would be convenient if users could determine
the scope of where a process is running, e.g. is it the initial
program or a job?  Is it running under a system instance or
flux job instance?

Solution: Support a new flux_get_process_scope() function to
determine if a process is running under the:
- system instance
- non-system instance and is the initial program
- non-system instance and is a flux

Fixes #3812