flux-top: use appropriate variable name
Problem: The function get_instance_attr_int() uses an internal
variable name "level".  This is probably a cut and paste from
when the function got the "instance level", before it was converted
to a generic function to convert an attribute to an int.

Solution: Rename the variable name to "val".