broker: add 'kary' topology plugin
Problem: the broker's topology class cannot be extended without
changing the API.

Change the topology_create() prototype to accept an optional URI argument.
The URI scheme selects a plugin.  Also add a flux_error_t parameter that
can be passed through to the plugin init function, to capture human-readable
errors.

Add a built-in 'kary' plugin that implements a complete, k-ary tree.
Retain the convention that k = 0 means that rank 0 is the parent of all
the other ranks.  Drop topology_set_kary().

Update unit tests and users.