testsuite: remove deprecated flux-resource status tests
Problem: flux-resource list and status will soon be updated to match
the -o, --format=NAME behavior of flux-jobs(1), which means the use
-vv to select a different output format will go away, as will the
`-o help` usage to get a list of supported formats. This will break
a couple tests in t2351-resource-status-input.t.

Remove tests that use these options so that the testsuite does not
break when the changes are made.