flux-resource: remove --verbose option from status and list
Problem: The flux-resource status and list --verbose option is used
to select a different format, but a better way to do this will be
via the -o, --format=NAME option which will be added in a later commit.

Remove the --verbose option from both the flux-resource status and
list subcommands. Drop the extended formats for now.