flux-resource: support named formats via UtilConfig
Problem: flux-resource does not support named formats like
flux-jobs(1).

Add a FluxResourceConfig subclass of flux.util.UtilConfig which is
used by the status, drain, and list subcommands to configure a set
of named formats. Add the `-o, --format=` and `-n, --no-header`
options to flux-resource drain to match the other commands. Update
description of `--format` help usage string.

Start with a set of builtin formats for now.