flux-resource: add --format=rlist to flux-resource list
Problem: There is no equivalent to `flux resource list -v` which used
to display the full "rlist" (rank list + resources) in place of the
nodelist for each line.

Add a named format `--format=rlist` which includes the rlist output
field in place of nodelist to replace the old `-v` option.