flux-resource: fix performance issue with large resource sets
Problem: The flux-resource list subcommand is very slow for large
resource sets. For example a cluster with 3000 nodes will take over
1 minute to display the resulting lines of output. This is due to
use of the union() ResourceSet method, which creates a new resource
set on each invocation. This results in the unnecessary creation
and destruction of increasingly large resource set objects as the
program iterates through all ranks, resulting in an exponential
scaling problem.

Replace the use of union() with add(), which modifies the resource
set in place and avoids all creation and destruction of resource sets
within the per-rank loop. This results in an almost 100x speedup of
`flux resource list`.

Fixes #4700