rlist: add rlist_add()
Problem: librlist doesn't have a set addition function which is like
rlist_union() but modifies its argument instead of returning a new set.

Add

 rlist_add (rl1, rl2);

which adds resources not already in rl2 to the resource set in
rl1. This function is like rlist_union() but modifies its argument
instead of returning a new set, and like rlist_append() but it is
not an error to have duplicate resources in rl2.