python: expose add() method for ResourceSet objects
Problem: The ResourceSet class (and its Rlist implementation) do not
expose an add() method, which is like append() but only adds resources
missing from the target resource set instead of raising an error.

Add an abstract add() method to the ResourceSetImplementation class,
and expose the add() method in the Rlist and ResourceSet classes so
that this method is available to users.