flux-job: support multiple jobids for kill, raise, and cancel
Problem: The flux-job kill, raise, and cancel commands only take a
single jobid, which means these commands must be used in a loop to
act on multiple jobids. This is not only more difficult to use on
the command line, but is slow since RPCs to kill/raise/cancel each
job are back to back.

Support multiple jobids on the command line for each of these flux-job
subcommands. For jobs that take an optional note or message on the
command line, this is done by halting processing of jobids when
encountering the first argument that doesn't parse as a jobid. The
remaining arguments are then taken as the note.

The string "--" may also be used to force separation between jobids
and any note, it will neither be interpreted as a jobid nor as part
of the note.

Fixes #4714