broker: clear job environment vars in rc scripts
Problem: FLUX_JOB_ID and other job environment variables remain
set in the initial program if Flux was launched by Flux.  This
sends the wrong message to programs that need to detect if they're
running in a job environment or not, since these variables refer to
the Flux broker's job environment, not that of the current process.

Clear the job-related environment variables in the the broker rc
script environment, where the inital program is considered "rc2".

Drop unsetenv() calls for FLUX_JOB_ID, FLUX_JOB_SIZE, FLUX_JOB_NNODES
in the broker mainline, now redundant.

Fixes #4716