flux-job: add --message option to cancel, raise subcommands
Problem: It may be inconvenient to require users to set the optional
excpetion note in the free arguments of the flux-job cancel and raise
subcommands, but there is currently no other way to do it.

Add a -m, --message option to flux-job cancel and raise which allows
the exception note to be set via a specific option.