testsuite: update tests for flux-job cancel/kill/raise
Problem: The flux-job cancel, kill, and raise subcommands can now
operate on multiple jobs at once, but this is not tested nor utilized
in the testsuite.

Add some tests of this feature, and the use of the -m, --message
option to t2201-job-cmd.t.

Additionally, replace some uses of `for id in ...; flux job cancel
$id..` with a single `flux job cancel ids...` call througout the
testsuite.