job-list: ensure purged jobs are inactive
Problem: The handler for the job-purge-inactive event does not check
if the requested job ids are inactive.  Purging a non-inactive job
id can lead to various issues such as stats miscalculations and
(worst case) an assertion could be triggered.

Add check to ensure job is inactive before doing purge.

Fixes #4711