add proto example using Python
problem: we want to eventually support a more standard rpc
interface, and using Captain Proto (cue superhero music)
can likely help us. This is not intended to merge in, but
is an example of creating a simple client/server model
using the Python library for captain proto, and it uses
a schema for a FluxMessage to create the message, and then
send to the server, and the server creates the flux RPC
and returns the result (serialized as a string, which
likely we can improve upon.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>