ci: run only some tests during system testing
Problem: The system coverage test takes a very long time, and many
of the tests being run are duplicates of the normal coverage build.

Add a `# ci=system` tag to tests that must be run in the system test
builder. Run only the tests with this tag in the system coverage build
in order to greatly speed up the build while still collecting coverage
for tests that aren't necessary run or are run differently in the
normal coverage build.