completion: use bash-completion init function if available
Problem: Handling of word splitting for bash completion appears to be
dependent on the bash version, contents of COMP_WORDBREAKS, and perhaps
even which other completions are loaded. Use of the bash-completion
project's _init_completion() function seems to better handle some of
these corner cases, but it isn't always guaranteed that bash-completion
will be available.

Check for _init_completion() and use it to set $cur and $prev
in the flux-core completion initialization function when it
available. Otherwise, fall back to the previous behavior. This should
result in more predictable behavior when bash-completion is available.