job-list: do not assume errno set on czmq funcs
Problem: A number of error logging functions assumed zhashx and
zlistx functions set errno and thus flux_log_error() was called to
log an error.  However it is not a safe assumption to assume they
set errno on error.

Solution: Replace flux_log_error() calls with flux_log() calls after
zhashx or zlistx errors.  When necessary, set errno = ENOMEM or remove
setting of errno = ENOMEM when it is not necessary to set.

Fixes #4741