job-list: remove excess logging calls
Problem: A number of internal functions log error messages that
are eventually logged via their callers.  Thus there is an excess
and double logging that is unnecessary.

Solution: Remove a number of excess logging of errors.  In a few
cases, log an error message at a caller so multiple log calls
can be removed from a callee.