testsuite: ensure job shell options are unset when expected
Problem: A test in t2603-job-shell-initrc.t assumes that
attributes.system.shell.options is unset for the test, but this may
no longer be true if some job shell options are set by default.

Use jq to remove attributes.system.shell.options if set in the jobspec
for the affected test.