job-list: add purge timestamp to job stats
Problem: There is no way to know if job stats have been modified
due to a purge.  This can be confusing to those reading / monitoring
stats since the numbers may not reflect what they are expecting.

Solution: Add an optional t_last_purge field to the job stats.  It will
indicate the last time a purge of job(s) has occurred in the stats.