github: use v1 for softprops/action-gh-release
Problem: In commit 286df0a057c7829bd5e842b457137e600c944a15 the
softprops/actions-gh-release was pinned to a "newer" release by git
hash, but this project seems to move the v1 tag to the latest release
in the v1.x series, so the version was actually downgraded. This came
with some bugs that have been fixed.

Revert the softprops/action-gh-release version to @v1.