shell: add unshare plugin
Problem: it would be nice if jobs could mount file systems
like a private /tmp or synthetic file systems.

Add a simplistic "unshare" shell plugin that can instruct the
shell to call unshare(2) with various flags, and also optionally
remap the user id in the job to root, which allows it to operate
with CAP_SYS_ADMIN and other capabilities within the namespace.