libtaskmap: add initial taskmap implementation
Problem: There is no helper library for decoding taskmap objects
as described in RFC 34.

Add an initial taskmap class implementation, which can encode and
decode RFC 34 taskmaps. Export this interface in a libflux-taskmap.so
library and flux/taskmap.h header.

Add a set of unit tests for the taskmap class.