python: add flux.job.get_job
Problem: most users of the Python API (or any API in
general) ideally want an intuitive user interface.
An early likely interaction is to do a flux.job.submit,
and then (for many use cases) the next desire is to get
more information about the job. As currently implemented,
the user needs to stumble on flux.job.job_list_id and
decide to try it, and then further to look at the rpc
object returned and figure out they can do get_str and
json loads that, OR a less intuitive "get" to do the same.
OR they could figure out some interaction with a JobInfo
object, which to be frank, I do not see how they would
know to go there because it is located in a totally
separate place, and requires knowing how to instantiate
(and is not just calling an easy function). See
https://github.com/flux-framework/flux-core/issues/4760

Solution: the ideal solution from the point of user
expectation is to place an obvious "get_job" directly
on the same module the user is already interacting with.
It can be quickly found and returns the expected data
structure without further parsing needed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>