shell: update and clean up old rcalc code
Problem: The code in src/shell/rcalc.c was written before libidset,
and thus has its own parsing of idset-style cpu and gpu ids from Rv1,
and uses the cpu_set_t type to store the cpuset and gpuset.

Update the code to use libidset while parsing Rv1 and store the set
of cpus and gpus as idsets.