testsuite: remove basis from rcalc tests
Problem: The shell 'rcalc' code will no longer be responsible for
assigning taskids, and thus the rcalc_rankfinfo 'basis' member will
be removed. This will break all tests in t2600-job-shell-rcalc.t.

Remove references to basis in the t/shell/rcalc test program as well
as the expected output files in preparation for this change.